Don't require specific version of cryptography library.
This partially reverts 12d6184713 that was supposer to fix OpenSSL bug
by requiring cryptography Python package that is built with newer
OpenSSL library.

The problem with this approach is that Linux distributions (like Fedora)
ship this package built with OpenSSL dynamically linked and update
OpenSSL independently.

Requiring newest cryptography package is not just unnecessary, but also
doesn't work for these distributions because they have older version of
cryptography library.

Signed-off-by: Oleg Girko <ol@infoserver.lv>